requirements.txt
anyio==3.6.2
bidict==0.22.1
click==8.1.3
h11==0.14.0
httptools==0.5.0
idna==3.4
py-bcrypt==0.4
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.8.0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
trueskill==0.4.5
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
